added intermediate upgrade group and (now unused) endpoint

+++ clients/android/app/src/main/java/app/lockbook/util/CoreData.kt
@@ -65,7 +65,7 @@ data class UsageMetrics(
 
 @Serializable
 data class UsageItemMetric(
-    val exact: Int,
+    val exact: Long,
     val readable: String,
 )
